test(kvpp): consolidate feature coverage and document the design
