+++ .github/workflows/configs/nightly_config.yaml
@@ -181,15 +181,6 @@ a3:
         size: 2
   single_node:
     test_config:
-      - name: kvpp-mtp
-        os: linux-aarch64-nightly-a3-16
-        tests: tests/e2e/nightly/single_node/test_kvpp.py::test_kvpp_mtp
-      - name: kvpp-indexer-c8
-        os: linux-aarch64-nightly-a3-16
-        tests: tests/e2e/nightly/single_node/test_kvpp.py::test_kvpp_indexer_c8
-      - name: kvpp-pp
-        os: linux-aarch64-nightly-a3-4
-        tests: tests/e2e/nightly/single_node/test_kvpp.py::test_kvpp_pp
       - name: mtpx-deepseek-r1-0528-w8a8
         os: linux-aarch64-nightly-a3-16
         config_file_path: MTPX-DeepSeek-R1-0528-W8A8.yaml
