+++ docs/source/developer_guide/Design_Documents/index.md
@@ -1,3 +1,5 @@
 # Design Documents
 
 This section provides an overview of the features implemented in vLLM Ascend. Developers can refer to this guide to understand how vLLM Ascend works.
+
+- [KVPP: KV Cache Layer Parallelism](kvpp.md) — Physical cache placement, full-layer broadcast, and test design.
