+++ mkdocs.yml
@@ -333,6 +333,7 @@ nav:
         - developer_guide/Design_Documents/model_runner_v2_eplb.md
         - developer_guide/Design_Documents/ACL_Graph.md
         - developer_guide/Design_Documents/KV_Cache_Pool_Guide.md
+        - developer_guide/Design_Documents/kvpp.md
         - developer_guide/Design_Documents/add_custom_aclnn_op.md
         - developer_guide/Design_Documents/context_parallel.md
         - developer_guide/Design_Documents/balance_schedule_refactor.md
